Update the pytest for option1 fileread
The docker and shining panda in Jenkins is running on Linux its file system is case sensitive. The unit test passed in Windows but in Shining Panda and Docker it failed. So I have to remove the test_filename_case_sensitive test case. And that the for test_filenameinput, I changed check_filename("maze.csv") to check_filename("Maze.csv")